change composeTo reducer to conversations module
