+++ packages/ringcentral-integration/modules/ComposeText/actionTypes.js
@@ -10,7 +10,5 @@ export default new Enum([
   'addToNumber',
   'removeToNumber',
   'updateMessageText',
-  'toNumberMatched',
-  'addEntity',
-  'removeEntity'
+  'toNumberMatched'
 ], 'composeText');
