+++ packages/ringcentral-integration/modules/ComposeText/index.js
@@ -1,4 +1,3 @@
-import messageDirection from 'ringcentral-integration/enums/messageDirection';
 import { Module } from '../../lib/di';
 import RcModule from '../../lib/RcModule';
 import isBlank from '../../lib/isBlank';
@@ -22,9 +21,8 @@ import proxify from '../../lib/proxy/proxify';
     'MessageSender',
     'NumberValidate',
     'RolesAndPermissions',
-    { dep: 'ContactMatcher', optional: true },
+    { dep: 'Conversations', optional: true },
     { dep: 'ContactSearch', optional: true },
-    { dep: 'ConversationLogger', optional: true },
     { dep: 'ComposeTextOptions', optional: true }
   ]
 })
@@ -47,8 +45,7 @@ export default class ComposeText extends RcModule {
     numberValidate,
     contactSearch,
     rolesAndPermissions,
-    contactMatcher,
-    conversationLogger,
+    conversations,
     ...options
   }) {
     super({
@@ -66,8 +63,7 @@ export default class ComposeText extends RcModule {
     this._messageSender = messageSender;
     this._numberValidate = numberValidate;
     this._contactSearch = contactSearch;
-    this._contactMatcher = contactMatcher;
-    this._conversationLogger = conversationLogger;
+    this._conversations = conversations;
     this._lastContactSearchResult = [];
     this.senderNumbersList = [];
     storage.registerReducer({ key: this._storageKey, reducer: this._cacheReducer });
@@ -218,57 +214,8 @@ export default class ComposeText extends RcModule {
         return null;
       }
     }
-    this.addEntitys(this.toNumbers);
+    this._conversations.addEntitys(this.toNumbers);
     return this._messageSender.send({ fromNumber, toNumbers, text });
-  }
-  get correspondentMatch() {
-    return this.state.correspondentMatch;
-  }
-  addEntitys(entitys) {
-    this.store.dispatch({
-      type: this.actionTypes.addEntity,
-      entitys,
-    });
-  }
-  removeEntity(entity) {
-    this.store.dispatch({
-      type: this.actionTypes.removeEntity,
-      entity
-    });
-  }
-  relateCorrespondentEntity(responses) {
-    if (!this._contactMatcher ||
-      !this._conversationLogger ||
-      !this.correspondentMatch.length) {
-      return;
-    }
-    responses.forEach((response) => {
-      const {
-        conversation: {
-          id
-        }
-      } = response;
-      const correspondentMatch = this.correspondentMatch;
-      const number = response.direction === messageDirection.inbound ? response.from : response.to;
-      if (number.length !== 1) {
-        return;
-      }
-      const phoneNumber = number[0].phoneNumber || number[0].extensionNumber;
-      const correspondentMatches = this._contactMatcher.dataMapping[phoneNumber];
-      const correspondentEntity = correspondentMatches.filter(match =>
-        (correspondentMatch.some(innerMatch => match.id === innerMatch.rawId)));
-      let entity = null;
-      if (correspondentEntity.length === 1) {
-        [entity] = correspondentEntity;
-        this.removeEntity(entity);
-      }
-      if (entity) {
-        this._conversationLogger.logConversation({
-          correspondentEntity: entity,
-          conversationId: id
-        });
-      }
-    });
   }
 
   @proxify
