+++ packages/ringcentral-integration/modules/Conversations/actionTypes.js
@@ -20,4 +20,6 @@ export default new Enum([
   'deleteConversation',
   'increaseCurrentPage',
   'resetCurrentPage',
+  'addEntity',
+  'removeEntity'
 ], 'conversations');
