+++ packages/ringcentral-integration/modules/Conversations/index.js
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect';
-
+import messageDirection from '../../enums/messageDirection';
 import RcModule from '../../lib/RcModule';
 import { Module } from '../../lib/di';
 import getter from '../../lib/getter';
@@ -878,5 +878,54 @@ export default class Conversations extends RcModule {
 
   get _hasPermission() {
     return this._rolesAndPermissions.hasReadMessagesPermission;
+  }
+  get correspondentMatch() {
+    return this.state.correspondentMatch;
+  }
+  addEntitys(entitys) {
+    this.store.dispatch({
+      type: this.actionTypes.addEntity,
+      entitys,
+    });
+  }
+  removeEntity(entity) {
+    this.store.dispatch({
+      type: this.actionTypes.removeEntity,
+      entity
+    });
+  }
+  relateCorrespondentEntity(responses) {
+    if (!this._contactMatcher ||
+      !this._conversationLogger ||
+      !this.correspondentMatch.length) {
+      return;
+    }
+    responses.forEach((response) => {
+      const {
+        conversation: {
+          id
+        }
+      } = response;
+      const correspondentMatch = this.correspondentMatch;
+      const number = response.direction === messageDirection.inbound ? response.from : response.to;
+      if (number.length !== 1) {
+        return;
+      }
+      const phoneNumber = number[0].phoneNumber || number[0].extensionNumber;
+      const correspondentMatches = this._contactMatcher.dataMapping[phoneNumber];
+      const correspondentEntity = correspondentMatches.filter(match =>
+        (correspondentMatch.some(innerMatch => match.id === innerMatch.rawId)));
+      let entity = null;
+      if (correspondentEntity.length === 1) {
+        [entity] = correspondentEntity;
+        this.removeEntity(entity);
+      }
+      if (entity) {
+        this._conversationLogger.logConversation({
+          correspondentEntity: entity,
+          conversationId: id
+        });
+      }
+    });
   }
 }
